requirements.txt
PyQt6
selenium
webdriver-manager
openai
python-dotenv
requests